yargi_gpt5_tavil_final_part_Final/requirements.txt
fastapi
uvicorn[standard]
PyPDF2
python-dotenv
openai
langchain-text-splitters
pinecone
tavily-python
requests
streamlit
python-docx
httpx
markitdown
pydantic
